+++ src/collective/behavior/talcondition/utils.py
@@ -9,7 +9,7 @@ logger = logging.getLogger('collective.behavior.talcondition')
 WRONG_TAL_CONDITION = "The TAL expression '{0}' for element at '{1}' is wrong.  Original exception : {2}"
 
 
-def evaluateExpressionFor(obj, extra_expr_ctx={}, error_pattern=WRONG_TAL_CONDITION):
+def evaluateExpressionFor(obj, extra_expr_ctx={}, error_pattern=WRONG_TAL_CONDITION, raise_on_error=False):
     """Evaluate the expression stored in 'tal_condition' of given p_obj.
     """
     # get tal_condition
@@ -23,7 +23,8 @@ def evaluateExpressionFor(obj, extra_expr_ctx={}, error_pattern=WRONG_TAL_CONDIT
                                expression=tal_condition,
                                roles_bypassing_expression=roles_bypassing_talcondition,
                                extra_expr_ctx=extra_expr_ctx,
-                               error_pattern=error_pattern)
+                               error_pattern=error_pattern,
+                               raise_on_error=raise_on_error)
 
 
 def _evaluateExpression(obj,
@@ -31,7 +32,8 @@ def _evaluateExpression(obj,
                         roles_bypassing_expression=[],
                         extra_expr_ctx={},
                         empty_expr_is_true=True,
-                        error_pattern=WRONG_TAL_CONDITION):
+                        error_pattern=WRONG_TAL_CONDITION,
+                        raise_on_error=False):
     """Evaluate given p_expression extending expression context with p_extra_expr_ctx."""
     if not expression or not expression.strip():
         return empty_expr_is_true
@@ -50,12 +52,16 @@ def _evaluateExpression(obj,
     ctx.setGlobal('portal', portal)
     for extra_key, extra_value in extra_expr_ctx.items():
         ctx.setGlobal(extra_key, extra_value)
-    try:
+
+    if raise_on_error:
         res = Expression(expression)(ctx)
-    except Exception, e:
-        logger.warn(error_pattern.format(
-            expression, obj.absolute_url(), str(e)))
-        res = False
+    else:
+        try:
+            res = Expression(expression)(ctx)
+        except Exception, e:
+            logger.warn(error_pattern.format(
+                expression, obj.absolute_url(), str(e)))
+            res = False
     return res